update last 10 transaction

+++ app/src/main/java/com/startwithn/exchange_android/model/response/TransactionsModel.kt
@@ -26,5 +26,5 @@ data class TransactionsModel(
     @field:SerializedName("type_text")
     var typeText: String?,
     @field:SerializedName("value")
-    var value: Int?
+    var value: Double?
 ) : Parcelable
